fix invalid left shifts

+++ tests/compression.c
@@ -222,7 +222,7 @@ main(int argc, char *argv[])
             for (i = 0; i < fdims[0]; i++)
             {
                 double x = 2 * M_PI * (double) i / (double) (fdims[0]-1);
-                double n = noise * ((double) random() / ((double)(1<<31)-1) - 0.5);
+                double n = noise * ((double) random() / ((double)((long long)1<<31)-1) - 0.5);
                 fval[i] = (float) ((j+1) * (1 + sin(x)) + n);
             }
 
@@ -254,7 +254,7 @@ main(int argc, char *argv[])
             for (i = 0; i < ddims[0]; i++)
             {
                 double x = 2 * M_PI * (double) i / (double) (ddims[0]-1);
-                double n = noise * ((double) random() / ((double)(1<<31)-1) - 0.5);
+                double n = noise * ((double) random() / ((double)((long long)1<<31)-1) - 0.5);
                 dval[i] = (double) ((j+1) * (1 + sin(x)) + n);
             }
 
@@ -337,7 +337,7 @@ main(int argc, char *argv[])
           for (i = 0; i < fdims[0]; i++)
           {
              double x = 2 * M_PI * (double) i / (double) (fdims[0]-1);
-             double n = noise * ((double) random() / ((double)(1<<31)-1) - 0.5);
+             double n = noise * ((double) random() / ((double)((long long)1<<31)-1) - 0.5);
              fval[i] = (float) ((j+1) * (1 + sin(x)) + n);
              if (fval[i] != frval[i])
              {
@@ -383,7 +383,7 @@ main(int argc, char *argv[])
           for (i = 0; i < ddims[0]; i++)
           {
              double x = 2 * M_PI * (double) i / (double) (ddims[0]-1);
-             double n = noise * ((double) random() / ((double)(1<<31)-1) - 0.5);
+             double n = noise * ((double) random() / ((double)((long long)1<<31)-1) - 0.5);
              dval[i] = (double) ((j+1) * (1 + sin(x)) + n);
              if (dval[i] != drval[i])
              {
